Added root check to startup daemon, termination if no root privileges are detected
Netatalk3 on macOS currently needs to be run with root privileges (in theory, its binaries can be executed within a normal user's context, but they won't provide any useful functionality). To reflect that, this four-liner checks if the startup daemon is executed with root privileges and will terminate the script if no root privileges are detected.